feat(settings): Hugging Face connect flow via a generic token-connect component
Generalizes the Modal connect card/dialog/form into config-driven
TokenConnect* components (provider = i18n namespace + token specs with
prefix patterns) and adds a second provider: Hugging Face. The HF_TOKEN
row in the Compute section becomes a connect card with the same guided
paste flow (story, huggingface.co/settings/tokens jump-off, one-shot
paste with hf_ extraction, masked connected state, disconnect).
