+++ src/i18n/messages/en.json
@@ -62,8 +62,8 @@
     "ModalConnect": {
         "title": "Connect Modal",
         "subtitle": "Run GPU nodes on your own compute in a couple of minutes.",
-        "storyRun": "TongFlow runs GPU workloads in your own Modal account — your billing, your data.",
-        "storyCredits": "Modal includes $30 of free compute every month.",
+        "story1": "TongFlow runs GPU workloads in your own Modal account — your billing, your data.",
+        "story2": "Modal includes $30 of free compute every month.",
         "storyManaged": "Your token stays yours — encrypted at rest.",
         "storyLocal": "Your token never leaves this machine.",
         "step1Title": "Step 1 — Grab a token from Modal",
@@ -92,6 +92,34 @@
         "detectedSecret": "Token Secret detected",
         "missingId": "Waiting for the Token ID (starts with ak-)",
         "missingSecret": "Waiting for the Token Secret (starts with as-)"
+    },
+    "HfConnect": {
+        "title": "Connect Hugging Face",
+        "subtitle": "Needed to download open model weights — takes a minute.",
+        "story1": "Open models like SAM, Wan and Flux host their weights on Hugging Face; downloading them needs your access token.",
+        "story2": "Signing up is free and a Read token is enough. A few models also ask you to accept their license on the model page.",
+        "storyManaged": "Your token stays yours — encrypted at rest.",
+        "storyLocal": "Your token never leaves this machine.",
+        "step1Title": "Step 1 — Grab a token from Hugging Face",
+        "openTokens": "Open huggingface.co/settings/tokens",
+        "signupHint": "No account yet? Signing up is free.",
+        "step2Title": "Step 2 — Paste it here",
+        "pasteHint": "Paste your access token (starts with hf_) — any surrounding text is fine.",
+        "detectedToken": "Token {id}",
+        "missingToken": "Waiting for the token (starts with hf_)",
+        "saveConnect": "Save & connect",
+        "connectedToast": "Hugging Face connected",
+        "connectFailed": "Couldn't save the token — please try again",
+        "cardTitle": "Hugging Face account",
+        "cardBlurb": "Connect to auto-download weights for open models like SAM and Wan — free, a Read token is enough.",
+        "connectCta": "Connect Hugging Face",
+        "connectedBadge": "Connected",
+        "reconnect": "Reconnect",
+        "disconnect": "Disconnect",
+        "disconnectedToast": "Hugging Face disconnected",
+        "disconnectConfirmTitle": "Disconnect Hugging Face?",
+        "disconnectConfirmBody": "Plugins that need Hugging Face weights won't be able to download models until you connect again.",
+        "disconnectCancel": "Cancel"
     },
     "Onboarding": {
         "welcomeTitle": "Welcome to TongFlow",
