+++ src/i18n/messages/ja.json
@@ -62,8 +62,8 @@
     "ModalConnect": {
         "title": "Modal を連携",
         "subtitle": "数分で自分のコンピュートで GPU ノードを動かせます。",
-        "storyRun": "TongFlow の GPU 処理はあなた自身の Modal アカウントで実行されます。請求もデータもあなたのものです。",
-        "storyCredits": "Modal では毎月 $30 分の無料コンピュートが使えます。",
+        "story1": "TongFlow の GPU 処理はあなた自身の Modal アカウントで実行されます。請求もデータもあなたのものです。",
+        "story2": "Modal では毎月 $30 分の無料コンピュートが使えます。",
         "storyManaged": "トークンは暗号化して保存され、あなただけのものです。",
         "storyLocal": "トークンはこの端末の外に出ません。",
         "step1Title": "ステップ 1：Modal でトークンを取得",
@@ -92,6 +92,34 @@
         "detectedSecret": "トークンシークレットを検出",
         "missingId": "トークン ID が見つかりません（ak- で始まります）",
         "missingSecret": "シークレットが見つかりません（as- で始まります）"
+    },
+    "HfConnect": {
+        "title": "Hugging Face を連携",
+        "subtitle": "オープンモデルの重みのダウンロードに必要です。1 分で完了します。",
+        "story1": "SAM や Wan、Flux などのオープンモデルの重みは Hugging Face にあり、ダウンロードにはアクセストークンが必要です。",
+        "story2": "登録は無料で、Read 権限のトークンで十分。一部のモデルはモデルページでライセンスへの同意も必要です。",
+        "storyManaged": "トークンは暗号化して保存され、あなただけのものです。",
+        "storyLocal": "トークンはこの端末の外に出ません。",
+        "step1Title": "ステップ 1：Hugging Face でトークンを取得",
+        "openTokens": "huggingface.co/settings/tokens を開く",
+        "signupHint": "アカウントがなければ、まず無料登録を。",
+        "step2Title": "ステップ 2：ここに貼り付け",
+        "pasteHint": "アクセストークン（hf_ で始まる）を貼り付けてください。前後に文字が混ざっていても自動で読み取ります。",
+        "detectedToken": "トークン {id}",
+        "missingToken": "トークンが見つかりません（hf_ で始まります）",
+        "saveConnect": "保存して連携",
+        "connectedToast": "Hugging Face を連携しました",
+        "connectFailed": "保存に失敗しました。もう一度お試しください",
+        "cardTitle": "Hugging Face アカウント",
+        "cardBlurb": "連携すると SAM や Wan などのオープンモデルの重みを自動ダウンロードできます。無料で、Read 権限で十分です。",
+        "connectCta": "Hugging Face を連携",
+        "connectedBadge": "連携済み",
+        "reconnect": "再連携",
+        "disconnect": "連携を解除",
+        "disconnectedToast": "Hugging Face の連携を解除しました",
+        "disconnectConfirmTitle": "Hugging Face の連携を解除しますか？",
+        "disconnectConfirmBody": "解除すると HF の重みが必要なプラグインはモデルをダウンロードできなくなります（再連携すれば戻ります）。",
+        "disconnectCancel": "キャンセル"
     },
     "Onboarding": {
         "welcomeTitle": "TongFlow へようこそ",
