+++ src/i18n/messages/ko.json
@@ -62,8 +62,8 @@
     "ModalConnect": {
         "title": "Modal 연결",
         "subtitle": "몇 분이면 내 컴퓨팅에서 GPU 노드를 실행할 수 있어요.",
-        "storyRun": "TongFlow의 GPU 작업은 내 Modal 계정에서 실행됩니다. 요금도 데이터도 모두 내 것이에요.",
-        "storyCredits": "Modal은 매달 $30의 무료 컴퓨팅을 제공합니다.",
+        "story1": "TongFlow의 GPU 작업은 내 Modal 계정에서 실행됩니다. 요금도 데이터도 모두 내 것이에요.",
+        "story2": "Modal은 매달 $30의 무료 컴퓨팅을 제공합니다.",
         "storyManaged": "토큰은 암호화되어 저장되며 나만 사용할 수 있습니다.",
         "storyLocal": "토큰은 이 컴퓨터 밖으로 나가지 않습니다.",
         "step1Title": "1단계 — Modal에서 토큰 발급",
@@ -92,6 +92,34 @@
         "detectedSecret": "토큰 시크릿 인식됨",
         "missingId": "토큰 ID를 아직 찾지 못했어요(ak-로 시작)",
         "missingSecret": "토큰 시크릿을 아직 찾지 못했어요(as-로 시작)"
+    },
+    "HfConnect": {
+        "title": "Hugging Face 연결",
+        "subtitle": "오픈 모델 가중치를 내려받는 데 필요해요. 1분이면 끝납니다.",
+        "story1": "SAM, Wan, Flux 같은 오픈 모델의 가중치는 Hugging Face에 있어서, 내려받으려면 액세스 토큰이 필요합니다.",
+        "story2": "가입은 무료이고 Read 권한 토큰이면 충분해요. 일부 모델은 모델 페이지에서 라이선스 동의도 필요합니다.",
+        "storyManaged": "토큰은 암호화되어 저장되며 나만 사용할 수 있습니다.",
+        "storyLocal": "토큰은 이 컴퓨터 밖으로 나가지 않습니다.",
+        "step1Title": "1단계 — Hugging Face에서 토큰 발급",
+        "openTokens": "huggingface.co/settings/tokens 열기",
+        "signupHint": "계정이 없다면 먼저 무료로 가입하세요.",
+        "step2Title": "2단계 — 여기에 붙여넣기",
+        "pasteHint": "액세스 토큰(hf_로 시작)을 붙여넣으세요. 다른 글자가 섞여 있어도 자동으로 인식합니다.",
+        "detectedToken": "토큰 {id}",
+        "missingToken": "토큰을 아직 찾지 못했어요(hf_로 시작)",
+        "saveConnect": "저장하고 연결",
+        "connectedToast": "Hugging Face가 연결되었습니다",
+        "connectFailed": "저장에 실패했어요. 다시 시도해 주세요",
+        "cardTitle": "Hugging Face 계정",
+        "cardBlurb": "연결하면 SAM, Wan 같은 오픈 모델 가중치를 자동으로 내려받아요. 무료이고 Read 권한이면 충분합니다.",
+        "connectCta": "Hugging Face 연결",
+        "connectedBadge": "연결됨",
+        "reconnect": "다시 연결",
+        "disconnect": "연결 해제",
+        "disconnectedToast": "Hugging Face 연결을 해제했습니다",
+        "disconnectConfirmTitle": "Hugging Face 연결을 해제할까요?",
+        "disconnectConfirmBody": "해제하면 HF 가중치가 필요한 플러그인이 모델을 내려받을 수 없습니다(다시 연결하면 복구됩니다).",
+        "disconnectCancel": "취소"
     },
     "Onboarding": {
         "welcomeTitle": "TongFlow에 오신 것을 환영합니다",
