+++ src/i18n/messages/zh.json
@@ -62,8 +62,8 @@
     "ModalConnect": {
         "title": "连接 Modal",
         "subtitle": "几分钟就能在自己的算力上跑 GPU 节点。",
-        "storyRun": "TongFlow 的 GPU 任务在你自己的 Modal 账号里运行，账单和数据都归你。",
-        "storyCredits": "Modal 每月赠送 $30 免费算力，日常使用基本够用。",
+        "story1": "TongFlow 的 GPU 任务在你自己的 Modal 账号里运行，账单和数据都归你。",
+        "story2": "Modal 每月赠送 $30 免费算力，日常使用基本够用。",
         "storyManaged": "Token 云端加密存储，只属于你。",
         "storyLocal": "Token 只保存在你自己的电脑上。",
         "step1Title": "第一步：去 Modal 获取 Token",
@@ -92,6 +92,34 @@
         "detectedSecret": "已识别 Token Secret",
         "missingId": "还没识别到 Token ID（ak- 开头）",
         "missingSecret": "还没识别到 Token Secret（as- 开头）"
+    },
+    "HfConnect": {
+        "title": "连接 Hugging Face",
+        "subtitle": "下载开源模型权重需要它，一分钟搞定。",
+        "story1": "SAM、Wan、Flux 这类开源模型的权重托管在 Hugging Face，下载时需要你的访问令牌。",
+        "story2": "注册免费，Read 权限的 Token 就够用；个别模型还需在其页面上点一下同意协议。",
+        "storyManaged": "Token 云端加密存储，只属于你。",
+        "storyLocal": "Token 只保存在你自己的电脑上。",
+        "step1Title": "第一步：去 Hugging Face 获取 Token",
+        "openTokens": "打开 huggingface.co/settings/tokens",
+        "signupHint": "还没有账号？先免费注册一个。",
+        "step2Title": "第二步：粘贴到这里",
+        "pasteHint": "粘贴你的访问令牌（hf_ 开头），带着别的文字一起粘也能识别。",
+        "detectedToken": "Token {id}",
+        "missingToken": "还没识别到 Token（hf_ 开头）",
+        "saveConnect": "保存并连接",
+        "connectedToast": "Hugging Face 已连接",
+        "connectFailed": "保存失败，请重试",
+        "cardTitle": "Hugging Face 账号",
+        "cardBlurb": "连接后可自动下载 SAM、Wan 等开源模型的权重——免费，Read 权限即可。",
+        "connectCta": "连接 Hugging Face",
+        "connectedBadge": "已连接",
+        "reconnect": "重新连接",
+        "disconnect": "断开连接",
+        "disconnectedToast": "已断开 Hugging Face",
+        "disconnectConfirmTitle": "断开 Hugging Face？",
+        "disconnectConfirmBody": "断开后需要 HF 权重的插件将无法下载模型，重新连接即可恢复。",
+        "disconnectCancel": "取消"
     },
     "Onboarding": {
         "welcomeTitle": "欢迎使用 TongFlow",
